initial question component test
